+++ packages/react/src/composite/Card/CHANGELOG.md
@@ -2,6 +2,17 @@
 
 The Card component allows us to group contents and related actions, making it easier to identify groups of information.
 
+## 2023-03-13 `3.0.0`
+
+### 🛠 Breaking changes
+
+- Changed the `backgroundColor` properties to `primary-surface`, `primary-surfaceHighlight`, `success-surface`, `success-surfaceHighlight`, `warning-surface`, `warning-surfaceHighlight`, `danger-surface`, `danger-surfaceHighlight`, `neutral-background`, `neutral-surface` and `neutral-surfaceHighlight`. ([#110](https://github.com/TiendaNube/nimbus-design-system/pull/110) by [@juniorconquista](https://github.com/juniorconquista))
+
+### 💡 Others
+
+- Refactored use of color tokens in internal components by removing `.` per `-`. ([#110](https://github.com/TiendaNube/nimbus-design-system/pull/110) by [@juniorconquista](https://github.com/juniorconquista))
+- Refactored style classes to have more semantic names. ([#110](https://github.com/TiendaNube/nimbus-design-system/pull/110) by [@juniorconquista](https://github.com/juniorconquista))
+
 ## 2023-02-24 `2.1.1`
 
 ### 💡 Others